mv InteractiveRequireLoginAndToS into login package
