+++ src/pkg/login/login.go
@@ -7,12 +7,15 @@ import (
 	"os"
 
 	"github.com/DefangLabs/defang/src/pkg/auth"
+	"github.com/DefangLabs/defang/src/pkg/cli"
 	"github.com/DefangLabs/defang/src/pkg/cli/client"
 	"github.com/DefangLabs/defang/src/pkg/cluster"
 	"github.com/DefangLabs/defang/src/pkg/dryrun"
 	"github.com/DefangLabs/defang/src/pkg/github"
 	"github.com/DefangLabs/defang/src/pkg/term"
+	"github.com/DefangLabs/defang/src/pkg/track"
 	defangv1 "github.com/DefangLabs/defang/src/protos/io/defang/v1"
+	"github.com/bufbuild/connect-go"
 )
 
 type LoginFlow = auth.LoginFlow
@@ -106,4 +109,42 @@ func NonInteractiveGitHubLogin(ctx context.Context, client client.FabricClient, 
 		return err
 	}
 	return cluster.SaveAccessToken(fabric, resp.AccessToken)
+}
+
+func InteractiveRequireLoginAndToS(ctx context.Context, fabric client.FabricClient, addr string) error {
+	var err error
+	if err = fabric.CheckLoginAndToS(ctx); err != nil {
+		// Login interactively now; only do this for authorization-related errors
+		if connect.CodeOf(err) == connect.CodeUnauthenticated {
+			term.Debug("Server error:", err)
+			term.Warn("Please log in to continue.")
+			term.ResetWarnings() // clear any previous warnings so we don't show them again
+
+			defer func() { track.Cmd(nil, "Login", P("reason", err)) }()
+			if err = InteractiveLogin(ctx, fabric, addr); err != nil {
+				return err
+			}
+
+			// Reconnect with the new token
+			if fabric, err = cli.Connect(ctx, addr); err != nil {
+				return err
+			}
+
+			if err = fabric.CheckLoginAndToS(ctx); err == nil { // recheck (new token = new user)
+				return nil // success
+			}
+		}
+
+		// Check if the user has agreed to the terms of service and show a prompt if needed
+		if connect.CodeOf(err) == connect.CodeFailedPrecondition {
+			term.Warn(client.PrettyError(err))
+
+			defer func() { track.Cmd(nil, "Terms", P("reason", err)) }()
+			if err = InteractiveAgreeToS(ctx, fabric); err != nil {
+				return err // fatal
+			}
+		}
+	}
+
+	return err
 }
